Versiyon 1.45 (2016.08.12 03:10) : - Ogrenciler Foto Yukleme
+ Okul logosu
+ takvim kisa ders adlari
+ barkodun turu
+ optik kodlama satir siralama
+ Www.ortaksinavmerkezi.com 》kullanıcı adı : #TC  şifre : #TC gibi metinler
+ Dağıtıma kitapçık türünü de koyarsak kitapçığıda kodlaması gerek
+ toplu fotograf yukleme ogrenci okul - sinif -noya gore olacak
+ takvim 6 derse gore
+ takvim seviyesini haftaya gore kaydetme
+ ayarlar default degerler
+ kodlama noktali degerler
+ ilk sayfada kaydirma

+++ src/main/java/net/yazsoft/ors/entities/OpticsParts.java
@@ -33,7 +33,7 @@ public class OpticsParts extends BaseEntity implements Serializable {
     private Date updated;
 
     private Integer rank;
-    private int x,y,w,h;
+    private float x,y,w,h;
     private boolean horizontal;
     private boolean print;
     private boolean printTitle;
@@ -114,35 +114,35 @@ public class OpticsParts extends BaseEntity implements Serializable {
         this.rank = rank;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public int getW() {
+    public float getW() {
         return w;
     }
 
-    public void setW(int w) {
+    public void setW(float w) {
         this.w = w;
     }
 
-    public int getH() {
+    public float getH() {
         return h;
     }
 
-    public void setH(int h) {
+    public void setH(float h) {
         this.h = h;
     }
